fullscreen/api/element-ready-check-not-allowed-cross-origin.sub.html

<!DOCTYPE html>
<title>Element ready check with no allowfullscreen attribute</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="../trusted-click.js"></script>
<iframe name="cross-origin-default"></iframe>
<script>
    promise_test(async (t) => {
        document.onfullscreenchange = t.unreached_func(
            "document fullscreenchange event"
        );
        document.onfullscreenerror = t.unreached_func(
            "document fullscreenerror event"
        );

        const iframe = document.querySelector("iframe");
        await new Promise((resolve) => {
            iframe.onload = resolve;
            iframe.src =
                "http://{{hosts[][]}}:{{ports[http][0]}}/fullscreen/api/resources/attempt-fullscreen.html";
        });
        // Trigger the child frame to report as soon as its content is loaded.
        await trusted_click();
        const data = await promiseMessage(iframe);
        // When a message is received from a child frame, ensure that the report
        // matches the expectations.
        assert_false(
            data.report.result,
            `${data.report.frame} frame entered fullscreen, but allowfullscreen was not set`
        );
    }, "Element ready check with no allowfullscreen attribute");
</script>
